+++ cogs/quiz.py
@@ -16,7 +16,10 @@ from copy import deepcopy
 
 import anyio
 import discord
+import discord.ext.commands
+import num2alpha as n2a
 
+import src.fractalrhomb_globals as frg
 import src.quiz as qb
 from src.fractalrhomb_globals import value_or_default
 
@@ -34,13 +37,14 @@ class LoadedQuizzes:
 _loaded_quizzes: LoadedQuizzes = LoadedQuizzes()
 
 
-class QuizCog(discord.Cog):
+class Quiz(discord.Cog):
 	"""Class defining the quiz cog."""
 
 	class QuizView(discord.ui.View):
 		"""View for quizzes."""
 
 		logger: logging.Logger
+		ping_user: bool
 		quiz: qb.Quiz
 		current_question: int
 		answered_questions: dict[int, int | None]
@@ -61,11 +65,14 @@ class QuizCog(discord.Cog):
 			logger: logging.Logger,
 			quiz: qb.Quiz,
 			quiz_message: discord.Interaction | discord.WebhookMessage,
-		) -> "QuizCog.QuizView":
+			*,
+			ping_user: bool,
+		) -> "Quiz.QuizView":
 			"""Initialize the quiz view, including copying the quiz and setting up buttons."""
 			self.logger = logger
+			self.ping_user = ping_user
 			self.quiz = deepcopy(quiz)
-			super().__init__(timeout=1800.0, disable_on_timeout=True)
+			super().__init__(timeout=1800.0)
 
 			self.current_question = 0
 			self.answered_questions = dict.fromkeys(range(len(self.quiz.questions)))
@@ -118,26 +125,58 @@ class QuizCog(discord.Cog):
 
 			asyncio.get_event_loop().create_task(self.refresh_view(quiz_message))
 
-		async def refresh_view(
-			self, interaction: discord.Interaction | discord.WebhookMessage
-		) -> None:
+		async def refresh_view(self, interaction: discord.Interaction) -> None:
 			"""Refresh the view to add new answers."""
 			self.logger.debug("Refreshing quiz view!")
 
 			self.clear_items()
 
 			selection_options = []
+			answer_texts = []
+			selected_answer = self.answered_questions[self.current_question]
 			for answer_num, answer in enumerate(
 				self.quiz.get_question_options(self.current_question)
 			):
+				answer_text = answer.format(n2a.number_to_alpha(answer_num + 1))
+				answer_texts.append(answer_text)
+				label_text = answer_text
+				if len(label_text) > frg.MAX_COMMON_INTERACTION_TEXT_LENGHT:
+					label_text = (
+						frg.split_message(
+							[label_text],
+							"",
+							frg.MAX_COMMON_INTERACTION_TEXT_LENGHT - len("..."),
+						)[0]
+						+ "..."
+					)
+
 				selection_options.append(
 					discord.SelectOption(
-						label=f"{answer_num + 1}.",
+						label=label_text,
 						value=str(answer_num),
-						description=answer.text,
 					)
 				)
-			self.selection = discord.ui.Select(row=0, options=selection_options)
+
+			placeholder = None
+			if selected_answer is not None:
+				placeholder = answer_texts[selected_answer]
+				if len(placeholder) > frg.MAX_COMMON_INTERACTION_TEXT_LENGHT:
+					placeholder = (
+						frg.split_message(
+							[placeholder],
+							"",
+							frg.MAX_COMMON_INTERACTION_TEXT_LENGHT - len("..."),
+						)[0]
+						+ "..."
+					)
+			elif self.current_question in self.answered_required_questions:
+				placeholder = "Pick an answer (required)"
+			else:
+				placeholder = "Pick an answer"
+
+			self.selection = discord.ui.Select(
+				row=0, placeholder=placeholder, options=selection_options
+			)
 			self.selection.callback = self.make_selection
 
 			self.add_item(self.selection)
@@ -173,11 +212,18 @@ class QuizCog(discord.Cog):
 				)
 				self.add_item(self.submit_button)
 
+			response = []
 			question = self.quiz.get_question(self.current_question)
-			response_text = question.format(self.current_question + 1)
-			selected_answer = self.answered_questions[self.current_question]
-			if selected_answer is not None:
-				response_text += f"\n\n(selected answer: {selected_answer + 1}. {question.options[selected_answer].format()})"
+
+			response.extend(
+				(question.format(self.current_question + 1), "\n".join(answer_texts))
+			)
+
+			response_text = "\n\n".join(response)
+
+			if self.ping_user:
+				response_text = f"<@!{interaction.user.id}>\n{response_text}"
+
 			await interaction.edit(content=response_text, view=self)
 
 		async def make_selection(self, interaction: discord.Interaction) -> None:
@@ -230,37 +276,68 @@ class QuizCog(discord.Cog):
 			self.logger.debug("Got results! %r", quiz_result)
 			result_message, result_image = self.quiz.format_result(quiz_result)
 
-			self.children[-1].style = discord.ButtonStyle.success
-			self.disable_all_items()
-			await interaction.edit(view=self)
+			if self.ping_user:
+				result_message = f"<@!{interaction.user.id}>\n{result_message}"
 
 			if result_image is not None:
-				await interaction.respond(
-					result_message, ephemeral=True, file=discord.File(result_image)
+				await interaction.edit(
+					content=result_message, view=None, file=discord.File(result_image)
 				)
 			else:
-				await interaction.respond(result_message, ephemeral=True)
+				await interaction.edit(content=result_message, view=None)
+
+		@staticmethod
+		async def interaction_check(interaction: discord.Interaction) -> bool:
+			"""Check if the user can interact."""
+			return (
+				interaction.user is not None
+				and interaction.user.id
+				== interaction.message.interaction_metadata.user.id
+			)
+
+		@staticmethod
+		async def on_check_failure(interaction: discord.Interaction) -> None:
+			"""Respond when the interaction check fails."""
+			await interaction.respond(
+				"only the command author can interact", ephemeral=True
+			)
 
 	class StartQuizView(discord.ui.View):
 		"""View for starting a quiz."""
 
-		def __init__(self, logger: logging.Logger, quiz: qb.Quiz) -> "QuizCog.QuizView":
+		def __init__(
+			self, logger: logging.Logger, quiz: qb.Quiz, *, ping_user: bool
+		) -> "Quiz.QuizView":
 			"""Initialize the view for starting a quiz."""
 			self.logger = logger
 			self.quiz = quiz
-			super().__init__(disable_on_timeout=True)
+			self.ping_user = ping_user
+			super().__init__()
 
 		@discord.ui.button(label="Start", style=discord.ButtonStyle.primary)
 		async def start_button_callback(
 			self, _button: discord.ui.Button, interaction: discord.Interaction
 		) -> None:
 			"""Start the quiz."""
-			quiz_message = await interaction.respond(
-				"initializing quiz...", ephemeral=True
-			)
-			QuizCog.QuizView(self.logger, self.quiz, quiz_message)
+			Quiz.QuizView(self.logger, self.quiz, interaction, ping_user=self.ping_user)
 
-	def __init__(self, bot: discord.Bot) -> "QuizCog":
+		@staticmethod
+		async def interaction_check(interaction: discord.Interaction) -> bool:
+			"""Check if the user can interact."""
+			return (
+				interaction.user is not None
+				and interaction.user.id
+				== interaction.message.interaction_metadata.user.id
+			)
+
+		@staticmethod
+		async def on_check_failure(interaction: discord.Interaction) -> None:
+			"""Respond when the interaction check fails."""
+			await interaction.respond(
+				"only the command author can interact", ephemeral=True
+			)
+
+	def __init__(self, bot: discord.Bot) -> "Quiz":
 		"""Initialize the cog."""
 		self.bot: discord.Bot = bot
 		self.logger = logging.getLogger("fractalrhomb.cogs.quiz")
@@ -324,30 +401,54 @@ class QuizCog(discord.Cog):
 		return list(_loaded_quizzes.quizzes.keys())
 
 	@quiz_group.command(name="start")
+	@discord.ext.commands.cooldown(3, 60.0, discord.ext.commands.BucketType.user)
 	@discord.option(
 		"name",
 		str,
 		description="The name of the quiz",
 		autocomplete=discord.utils.basic_autocomplete(available_quizzes),
 	)
-	async def start_quiz(self, ctx: discord.ApplicationContext, name: str) -> None:
+	@discord.option(
+		"name",
+		str,
+		description="The name of the quiz",
+		autocomplete=discord.utils.basic_autocomplete(available_quizzes),
+	)
+	@discord.option(
+		"ephemeral",
+		bool,
+		description="Only shown to you (default: True)",
+	)
+	async def start_quiz(
+		self, ctx: discord.ApplicationContext, name: str, *, ephemeral: bool = True
+	) -> None:
 		"""Start a quiz."""
 		self.logger.info("Start quiz command used")
 
-		await ctx.defer(ephemeral=True)
+		if not ephemeral and not await frg.bot_channel_warning(ctx):
+			return
+
+		deferred = False
+		if not ctx.response.is_done():
+			await ctx.defer(ephemeral=ephemeral)
+			deferred = True
 
 		quiz = _loaded_quizzes.quizzes.get(name)
 
 		if quiz is None:
-			await ctx.respond("Not a valid quiz")
+			await ctx.respond("Not a valid quiz", ephemeral=ephemeral)
 			return
 
 		msg = quiz.format_title()
-		view = self.StartQuizView(self.logger, quiz)
+		view = self.StartQuizView(self.logger, quiz, ping_user=(not deferred))
 
-		await ctx.respond(msg, view=view, ephemeral=True)
+		if deferred:
+			await ctx.respond(msg, view=view, ephemeral=ephemeral)
+		else:
+			msg = f"<@!{ctx.author.id}>\n{msg}"
+			await ctx.respond(msg, view=view)
 
 
 def setup(bot: discord.Bot) -> None:
 	"""Set up the cog."""
-	bot.add_cog(QuizCog(bot))
+	bot.add_cog(Quiz(bot))